Fix #83 

After:
c:\git\AqualinkD>grep finished stdout-20190807_1640
Debug:  Thread 9,0xb5800f90 (Init PDA) finished in 16.191 sec
Debug:  Thread 14,0xb5800f50 (PDA init after wake) finished in 8.702 sec
Debug:  Thread 12,0xb5800f10 (Switch PDA device on/off) finished in 2.889 sec
Debug:  Thread 4,0xb5800ed0 (Set Spa heater setpoint) finished in 10.530 sec
Debug:  Thread 4,0xb5800e90 (Set Spa heater setpoint) finished in 3.013 sec
Debug:  Thread 4,0xb5800e50 (Set Spa heater setpoint) finished in 3.013 sec
Debug:  Thread 12,0xb5800dd0 (Switch PDA device on/off) finished in 5.446 sec
Debug:  Thread 12,0xb5800d90 (Switch PDA device on/off) finished in 2.361 sec
Debug:  Thread 12,0xb5800d50 (Switch PDA device on/off) finished in 0.869 sec
Debug:  Thread 12,0xb5800d10 (Switch PDA device on/off) finished in 0.136 sec
Debug:  Thread 12,0xb5800cd0 (Switch PDA device on/off) finished in 0.789 sec
Debug:  Thread 12,0xb5800c90 (Switch PDA device on/off) finished in 9.749 sec
Debug:  Thread 12,0xb5800c50 (Switch PDA device on/off) finished in 0.477 sec
Debug:  Thread 14,0xb5800c10 (PDA init after wake) finished in 3.668 sec
Debug:  Thread 14,0xb5800bd0 (PDA init after wake) finished in 3.703 sec
Debug:  Thread 14,0xb5800b90 (PDA init after wake) finished in 3.613 sec
Debug:  Thread 14,0xb5800b50 (PDA init after wake) finished in 3.721 sec